File>Open works
The menu bar is controlled from AppDelegate.swift. Previously this could not access GlobalVars, and so the view wouldn’t update after loading a video on the backend. By setting a GlobalVars instance in AppDelegate (and passing this to ContentView, rather than GlobalVars()) I have solved this problem.

I have also put vp and frames back in GlobalVars because global variables are bad and I don’t know what I was thinking…